agent/requirements.txt
mlflow>=2.20,<3
databricks-agents>=0.10,<1
databricks-ai-bridge>=0.1,<1
databricks-vectorsearch>=0.40,<1
databricks-sdk>=0.40,<1
pydantic>=2.6,<3
